refactor(server): merge state module into dependencies package
