+++ api/chatbot/dependencies/db.py
@@ -2,9 +2,24 @@ from collections.abc import AsyncGenerator
 from typing import Annotated
 
 from fastapi import Depends
-from sqlalchemy.ext.asyncio import AsyncSession
+from sqlalchemy.ext.asyncio import AsyncSession, create_async_engine
+from sqlalchemy.orm import sessionmaker
+from sqlalchemy.pool import NullPool
 
-from chatbot.state import sqlalchemy_ro_session, sqlalchemy_session
+from chatbot.config import settings
+
+
+sqlalchemy_engine = create_async_engine(
+    str(settings.postgres_primary_url),
+    poolclass=NullPool,
+)
+sqlalchemy_session = sessionmaker(
+    sqlalchemy_engine,
+    autocommit=False,
+    expire_on_commit=False,
+    autoflush=False,
+    class_=AsyncSession,
+)
 
 
 async def get_sqlalchemy_session() -> AsyncGenerator[AsyncSession, None]:
@@ -13,6 +28,19 @@ async def get_sqlalchemy_session() -> AsyncGenerator[AsyncSession, None]:
 
 
 SqlalchemySessionDep = Annotated[AsyncSession, Depends(get_sqlalchemy_session)]
+
+
+sqlalchemy_ro_engine = create_async_engine(
+    str(settings.postgres_standby_url),
+    poolclass=NullPool,
+)
+sqlalchemy_ro_session = sessionmaker(
+    sqlalchemy_ro_engine,
+    autocommit=False,
+    expire_on_commit=False,
+    autoflush=False,
+    class_=AsyncSession,
+)
 
 
 async def get_sqlalchemy_ro_session() -> AsyncGenerator[AsyncSession, None]:
