+++ api/chatbot/main.py
@@ -17,6 +17,7 @@ from starlette.routing import Mount
 
 from chatbot.config import settings
 from chatbot.dependencies import EmailHeaderDep, UserIdHeaderDep, UsernameHeaderDep
+from chatbot.dependencies.db import sqlalchemy_engine
 from chatbot.models import Base
 from chatbot.routers.chat import router as chat_router
 from chatbot.routers.conversation import router as conversation_router
@@ -24,7 +25,6 @@ from chatbot.routers.files import router as files_router
 from chatbot.routers.message import router as message_router
 from chatbot.routers.share import router as share_router
 from chatbot.schemas import UserProfile
-from chatbot.state import sqlalchemy_engine
 
 
 logger = logging.getLogger(__name__)
